done front office

+++ src/main/java/com/mg/dao/ReservationDAO.java
@@ -90,7 +90,7 @@ public class ReservationDAO implements GenericDAO<Reservation> {
         }
     }
 
-    public boolean cancelReservation(Long reservationId) {
+    public boolean cancelReservation(Integer reservationId) {
         Session session = null;
         try {
             session = HibernateUtil.getSessionFactory().openSession();
